% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking.R
\name{convert_hash_code}
\alias{convert_hash_code}
\title{Convert a submitted hash code to a learnr document
for marking}
\usage{
convert_hash_code(
  hash_code,
  assignment_path,
  rmd_path = tempfile(fileext = ".Rmd"),
  student = "Test student",
  run = TRUE
)
}
\arguments{
\item{hash_code}{Character. hash code to convert}

\item{assignment_path}{Path to assignment .Rmd}
}
\description{
Convert a submitted hash code to a learnr document
for marking
}
