% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fofpack_updated.R
\name{fofpack_up_to_date}
\alias{fofpack_up_to_date}
\title{Check if fofpack is up to date.}
\usage{
fofpack_up_to_date()
}
\value{
Invisibly returns \code{TRUE} if fofpack is up to date, \code{FALSE} otherwise
}
\description{
Always make sure fofpack is up to date before starting any assignment!
}
\examples{
fofpack_up_to_date()
}
