% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fofpack-package.R
\docType{package}
\name{sdmpack-package}
\alias{sdmpack}
\alias{sdmpack-package}
\title{sdmpack: FIU SDM Course Package}
\description{
Course material for FIU course on SDM
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rdinnager.github.io/sdmpack/}
}

}
\author{
\strong{Maintainer}: Russell Dinnage \email{rdinnage@fiu.edu} (\href{https://orcid.org/0000-0003-0846-2819}{ORCID})

}
\keyword{internal}
