% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{send_env_to_RStudio}
\alias{send_env_to_RStudio}
\title{Send an environment to RStudio Session}
\usage{
send_env_to_RStudio(env_from, remove = "dput_to_string")
}
\arguments{
\item{env_from}{The environment to send}
}
\value{
Nothing
}
\description{
Send an environment to RStudio Session
}
