% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{set_env}
\alias{set_env}
\title{Synchronise an environment with another}
\usage{
set_env(env_from, env_to = rlang::global_env())
}
\arguments{
\item{env_from}{The environment to synchronise with}

\item{env_to}{The environment to synchronise}
}
\value{
Nothing
}
\description{
This function takes everything in an environment and puts it
in another, replacing any existing object with the same names.
}
