% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\docType{data}
\name{submission_server}
\alias{submission_server}
\alias{submission_ui}
\title{Tutorial submission functions}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
submission_server(input, output)

submission_ui
}
\description{
The following functions have been modified from
code in the github repo https://github.com/mattblackwell/qsslearnr
Versions of this appear to have proliferated around github and I
have no idea who originally wrote this submission system.
If you happen to know, please let me know. All versions of this
begin with this preamble I reproduce as well:
The following function has modified from Colin
Rundel's learnrhash package, available at
https://github.com/rundel/learnrhash. Many thanks to Professor Rundel, who
has developed a fantastic tool for courses that teach R and use the learnr
package.

This note is also modified from Professor Rundel's description: Note that when
including these functions in a learnr Rmd document it is necessary that the
server function, \code{submission_server()}, be included in an R chunk where
\code{context="server"}. Conversely, any of the ui functions, \verb{*_ui()}, must \emph{not}
be included in an R chunk with a \code{context}.
}
\keyword{datasets}
